Added health and object pooling
added a health feature with a damage over time coroutine, added object pooling that created 10 inactive health packs but i haven't done anything that "spawns" them in and instead am using manually placed prefabs for now to test the health system. I have made it so the player loses 60 health over 3 seconds at the start.

Bugs:
messed up when it comes to the multiplayer aspect so I'll need someone to convert this all into "multiplayer code"

The 3 seconds of damage happen right at the start when the tester is still clicking host or client