Added a way to trigger the wind effect
Go fast enough and the wind effect triggers.
Also made a prefab folder.
Planning on adding a brief cooldown (5 seconds) to the speed effect reappearing and a brief period (2 seconds) for the player to regain speed and keep the effect from disappearing to prevent the speed effect from flickering when the player is close to the speed threshold